moved the helper functions to the helper class
