+++ gvmtools/helper.py
@@ -18,7 +18,9 @@
 
 import getpass
 import os
-import sys
+
+from sys import stderr
+from lxml import etree
 
 from gvm.errors import GvmError
 from gvm.xml import pretty_print
@@ -86,6 +88,51 @@ class Table:
         return "\n".join(row_strings)
 
 
+def yes_or_no(question):
+    """Asks the user to proceed or not in a gvmtools script
+
+    Arguments:
+        question (str): The condition the user should answer
+    """
+    reply = str(input(question + ' (y/n): ')).lower().strip()
+    if reply[0] == ('y'):
+        return True
+    if reply[0] == ('n'):
+        return False
+    else:
+        return yes_or_no("Please enter 'y' or 'n'")
+
+
+def error_and_exit(msg):
+    """Prints an error message and quits the gvmtools script
+
+    Arguments:
+        msg (str): The error message, that will be printed
+    """
+    print("\nError: {}\n".format(msg), file=stderr)
+    quit(1)
+
+
+def create_xml_tree(xml_doc):
+    """Creates an XML tree that can be read by an gvmtools script
+
+    Arguments:
+        xml_doc (str): Path to the xml document
+    """
+    try:
+        xml_tree = etree.parse(xml_doc)
+        xml_tree = xml_tree.getroot()
+    except IOError as err:
+        error_and_exit("Failed to read xml_file: {} (exit)".format(str(err)))
+    except etree.Error as err:
+        error_and_exit("Failed to parse xml_file: {} (exit)".format(str(err)))
+
+    if len(xml_tree) == 0:
+        error_and_exit("XML file is empty (exit)")
+
+    return xml_tree
+
+
 def do_not_run_as_root():
     if hasattr(os, 'geteuid') and os.geteuid() == 0:
         raise RuntimeError('This tool MUST NOT be run as root user.')
@@ -139,7 +186,7 @@ def run_script(path, global_vars):
     try:
         file = open(path, 'r', newline='').read()
     except FileNotFoundError:
-        print('Script {path} does not exist'.format(path=path), file=sys.stderr)
-        sys.exit(2)
+        print('Script {path} does not exist'.format(path=path), file=stderr)
+        quit(2)
 
     exec(file, global_vars)  # pylint: disable=exec-used
